+++ components/map/basic/src/leaflet/shapes/Polygons.js
@@ -14,7 +14,14 @@ export default class SearchMapPolygons {
 
   BASE_CLASSNAME = 'scm-map__area'
 
-  constructor({hoverStyles, onLayerClick, onPolygonWithBounds, showLabels}) {
+  constructor({
+    currentGeoCode,
+    hoverStyles,
+    onLayerClick,
+    onPolygonWithBounds,
+    showLabels
+  }) {
+    this.currentGeoCode = currentGeoCode
     this.hoverStyles = hoverStyles
     this.onLayerClick = onLayerClick
     this.onPolygonWithBounds = onPolygonWithBounds
@@ -35,15 +42,23 @@ export default class SearchMapPolygons {
   printPolygonOnMap({fitBound, map, polygon}) {
     const {BASE_CLASSNAME, hoverStyles} = this
     const className = fitBound ? `${BASE_CLASSNAME} fitBound` : BASE_CLASSNAME
+
+    const getTooltipElement = feature => feature.getTooltip()?.getElement()
     const polygonGeoJSon = L.geoJson(polygon, {
       className,
       onEachFeature: (feature, layer) => {
         layer.on({
           mouseout: function() {
             polygonGeoJSon.resetStyle(this)
+
+            const tooltipElement = getTooltipElement(this)
+            if (tooltipElement) tooltipElement.classList.remove('is-hover')
           },
           mouseover: function() {
             this.setStyle(hoverStyles)
+
+            const tooltipElement = getTooltipElement(this)
+            if (tooltipElement) tooltipElement.classList.add('is-hover')
 
             if (!L.Browser.ie && !L.Browser.opera) {
               this.bringToFront()
@@ -59,16 +74,26 @@ export default class SearchMapPolygons {
     polygonGeoJSon.addTo(map)
 
     if (this.showLabels) {
+      const that = this
+
       polygonGeoJSon.eachLayer(function(layer) {
-        if (!layer?.feature?.properties?.LocationName) return
+        if (
+          !layer?.feature?.properties?.LocationName ||
+          !layer?.feature?.properties?.Code
+        )
+          return
 
         try {
-          layer
-            .bindTooltip(layer.feature.properties.LocationName, {
-              permanent: true,
-              direction: 'center'
-            })
-            .openTooltip()
+          const {Code, LocationName} = layer.feature.properties
+
+          that.currentGeoCode !== Code &&
+            layer
+              .bindTooltip(LocationName, {
+                permanent: true,
+                direction: 'center',
+                className: fitBound ? 'is-fit-bound' : ''
+              })
+              .openTooltip()
         } catch (error) {}
       })
     }